Added an auto-scroller on messages sent
Modified the useEffect logic in push context

Managed to make the fast message update on send with "updateAfterSend()" method, JSONS are the same but listener still detects that they changed and rerenders.

I will remove this method from the context and just rerender everything after the message gets sent.

A bit slower but works fine. I leave the data in this commit just in case.